Added modified version of the arm obstacle avoidance simulation.
Most of the code comes from the great original simulation.
In this version, both the work space and joint space visualization are
done in the same figure, which provides a better understanding of the
simulation.

Obstacles have been made more tricky in the joint space.

Added facility to stop the simulatino by pressing 'q'.

Added some comments.